+++ src/gazebo_mavlink_interface.cpp
@@ -66,47 +66,25 @@ GazeboMavlinkInterface::GazeboMavlinkInterface() :
     zero_position_armed_ {},
     input_index_ {},
     mag_updated_(false),
-    baro_updated_(false),
-    diff_press_updated_(false),
     groundtruth_lat_rad(0.0),
     groundtruth_lon_rad(0.0),
     groundtruth_altitude(0.0),
-    mavlink_udp_port_(kDefaultMavlinkUdpPort),
-    mavlink_tcp_port_(kDefaultMavlinkTcpPort),
-    simulator_socket_fd_(0),
-    simulator_tcp_client_fd_(0),
-    use_tcp_(false),
-    qgc_udp_port_(kDefaultQGCUdpPort),
-    sdk_udp_port_(kDefaultSDKUdpPort),
-    remote_qgc_addr_ {},
-    local_qgc_addr_ {},
-    remote_sdk_addr_ {},
-    local_sdk_addr_ {},
-    qgc_socket_fd_(0),
-    sdk_socket_fd_(0),
-    serial_enabled_(false),
-    tx_q {},
-    rx_buf {},
-    m_status {},
-    m_buffer {},
-    io_service(),
-    serial_dev(io_service),
-    device_(kDefaultDevice),
-    baudrate_(kDefaultBaudRate),
     hil_mode_(false),
     hil_state_level_(false)
-    {}
+    {
+      mavlink_interface_ = std::make_shared<MavlinkInterface>();
+    }
 
 GazeboMavlinkInterface::~GazeboMavlinkInterface() {
+  mavlink_interface_->close();
+  // sigIntConnection_->~Connection();
+  // updateConnection_->~Connection();
 }
 
 void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
       const std::shared_ptr<const sdf::Element> &_sdf,
       ignition::gazebo::EntityComponentManager &_ecm,
       ignition::gazebo::EventManager &/*_eventMgr*/) {
-
-//   model_ = _model;
-//   world_ = model_->GetWorld();
 
   namespace_.clear();
   if (_sdf->HasElement("robotNamespace")) {
@@ -138,6 +116,7 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
   input_reference_.resize(n_out_max);
   // joints_.resize(n_out_max);
   // pids_.resize(n_out_max);
+  // joint_max_errors_.resize(n_out_max);
 //   for (int i = 0; i < n_out_max; ++i)
 //   {
 //     pids_[i].Init(0, 0, 0, 0, 0, 0, 0);
@@ -190,21 +169,6 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
 //           {
 //             std::string joint_name = channel->Get<std::string>("joint_name");
 //             joints_[index] = model_->GetJoint(joint_name);
-//             if (joints_[index] == nullptr)
-//             {
-//               gzwarn << "joint [" << joint_name << "] not found for channel["
-//                     << index << "] no joint control for this channel.\n";
-//             }
-//             else
-//             {
-//               gzdbg << "joint [" << joint_name << "] found for channel["
-//                     << index << "] joint control active for this channel.\n";
-//             }
-//           }
-//           else
-//           {
-//             gzdbg << "<joint_name> not found for channel[" << index
-//                   << "] no joint control will be performed for this channel.\n";
 //           }
 
 //           // setup joint control pid to control joint
@@ -232,6 +196,9 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
 //             double cmdMin = 0;
 //             if (pid->HasElement("cmdMin"))
 //               cmdMin = pid->Get<double>("cmdMin");
+//             if (pid->HasElement("errMax")) {
+//               joint_max_errors_[index] = pid->Get<double>("errMax");
+//             }
 //             pids_[index].Init(p, i, d, iMax, iMin, cmdMax, cmdMin);
 //           }
 //         }
@@ -252,29 +219,37 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
   if(_sdf->HasElement("hil_mode"))
   {
     hil_mode_ = _sdf->Get<bool>("hil_mode");
+    mavlink_interface_->SetHILMode(hil_mode_);
   }
 
   if(_sdf->HasElement("hil_state_level"))
   {
     hil_state_level_ = _sdf->Get<bool>("hil_state_level");
+    mavlink_interface_->SetHILStateLevel(hil_state_level_);
   }
 
+  bool serial_enabled=false;
   if(_sdf->HasElement("serialEnabled"))
   {
-    serial_enabled_ = _sdf->Get<bool>("serialEnabled");
+    serial_enabled = _sdf->Get<bool>("serialEnabled");
+    mavlink_interface_->SetSerialEnabled(serial_enabled);
   }
 
-  if (!serial_enabled_ && _sdf->HasElement("use_tcp"))
+  bool use_tcp = false;
+  if (!serial_enabled && _sdf->HasElement("use_tcp"))
   {
-    use_tcp_ = _sdf->Get<bool>("use_tcp");
+    use_tcp = _sdf->Get<bool>("use_tcp");
+    mavlink_interface_->SetUseTcp(use_tcp);
   }
-  ignmsg << "Connecting to PX4 SITL using " << (serial_enabled_ ? "serial" : (use_tcp_ ? "TCP" : "UDP")) << "\n";
+  ignmsg << "Connecting to PX4 SITL using " << (serial_enabled ? "serial" : (use_tcp ? "TCP" : "UDP")) << "\n";
 
   if (!hil_mode_ && _sdf->HasElement("enable_lockstep"))
   {
     enable_lockstep_ = _sdf->Get<bool>("enable_lockstep");
+    mavlink_interface_->SetEnableLockstep(enable_lockstep_);
   }
   ignmsg << "Lockstep is " << (enable_lockstep_ ? "enabled" : "disabled") << "\n";
+
 
   // When running in lockstep, we can run the simulation slower or faster than
   // realtime. The speed can be set using the env variable PX4_SIM_SPEED_FACTOR.
@@ -379,15 +354,10 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
     imu_update_interval_ = 1 / _sdf->Get<int>("imu_rate");
   }
 
-  mavlink_addr_ = htonl(INADDR_ANY);
   if (_sdf->HasElement("mavlink_addr")) {
     std::string mavlink_addr_str = _sdf->Get<std::string>("mavlink_addr");
     if (mavlink_addr_str != "INADDR_ANY") {
-      mavlink_addr_ = inet_addr(mavlink_addr_str.c_str());
-      if (mavlink_addr_ == INADDR_NONE) {
-        ignerr << "Invalid mavlink_addr: " << mavlink_addr_str << ", aborting\n";
-        abort();
-      }
+      mavlink_interface_->SetMavlinkAddr(mavlink_addr_str);
     }
   }
 
@@ -397,212 +367,53 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
 //   auto worldName = world_->GetName();
 // #endif
 
+  int mavlink_udp_port;
   if (_sdf->HasElement("mavlink_udp_port")) {
-    mavlink_udp_port_ = _sdf->Get<int>("mavlink_udp_port");
+    mavlink_udp_port = _sdf->Get<int>("mavlink_udp_port");
+    //TODO: Set Malvink UDP port
   }
-//   model_param(worldName, model_->GetName(), "mavlink_udp_port", mavlink_udp_port_);
+  // model_param(worldName, model_->GetName(), "mavlink_udp_port", mavlink_udp_port);
+  mavlink_interface_->SetMavlinkUdpPort(mavlink_udp_port);
 
+  int mavlink_tcp_port;
   if (_sdf->HasElement("mavlink_tcp_port")) {
-    mavlink_tcp_port_ = _sdf->Get<int>("mavlink_tcp_port");
+    mavlink_tcp_port = _sdf->Get<int>("mavlink_tcp_port");
   }
-//   model_param(worldName, model_->GetName(), "mavlink_tcp_port", mavlink_tcp_port_);
+  // model_param(worldName, model_->GetName(), "mavlink_tcp_port", mavlink_tcp_port);
+  mavlink_interface_->SetMavlinkTcpPort(mavlink_tcp_port);
 
-  local_qgc_addr_.sin_port = 0;
   if (_sdf->HasElement("qgc_addr")) {
     std::string qgc_addr = _sdf->Get<std::string>("qgc_addr");
     if (qgc_addr != "INADDR_ANY") {
-      local_qgc_addr_.sin_port = inet_addr(qgc_addr.c_str());
-      if (local_qgc_addr_.sin_port == 0) {
-        ignerr << "Invalid qgc_addr: " << qgc_addr << ", aborting\n";
-        abort();
-      }
+      mavlink_interface_->SetQgcAddr(qgc_addr);
     }
   }
   if (_sdf->HasElement("qgc_udp_port")) {
-    qgc_udp_port_ = _sdf->Get<int>("qgc_udp_port");
+    int qgc_udp_port = _sdf->Get<int>("qgc_udp_port");
+    mavlink_interface_->SetQgcUdpPort(qgc_udp_port);
   }
 
-  local_sdk_addr_.sin_port = 0;
   if (_sdf->HasElement("sdk_addr")) {
     std::string sdk_addr = _sdf->Get<std::string>("sdk_addr");
     if (sdk_addr != "INADDR_ANY") {
-      local_sdk_addr_.sin_port = inet_addr(sdk_addr.c_str());
-      if (local_sdk_addr_.sin_port == 0) {
-        ignerr << "Invalid sdk_addr: " << sdk_addr << ", aborting\n";
-        abort();
-      }
+      mavlink_interface_->SetSdkAddr(sdk_addr);
     }
   }
   if (_sdf->HasElement("sdk_udp_port")) {
-    sdk_udp_port_ = _sdf->Get<int>("sdk_udp_port");
-  }
-
-  if (hil_mode_) {
-
-    local_qgc_addr_.sin_family = AF_INET;
-    local_qgc_addr_.sin_port = htons(0);
-    local_qgc_addr_len_ = sizeof(local_qgc_addr_);
-
-    remote_qgc_addr_.sin_family = AF_INET;
-    remote_qgc_addr_.sin_port = htons(qgc_udp_port_);
-    remote_qgc_addr_len_ = sizeof(remote_qgc_addr_);
-
-    local_sdk_addr_.sin_family = AF_INET;
-    local_sdk_addr_.sin_port = htons(0);
-    local_sdk_addr_len_ = sizeof(local_sdk_addr_);
-
-    remote_sdk_addr_.sin_family = AF_INET;
-    remote_sdk_addr_.sin_port = htons(sdk_udp_port_);
-    remote_sdk_addr_len_ = sizeof(remote_sdk_addr_);
-
-    if ((qgc_socket_fd_ = socket(AF_INET, SOCK_DGRAM, 0)) < 0) {
-      ignerr << "Creating QGC UDP socket failed: " << strerror(errno) << ", aborting\n";
-      abort();
-    }
-
-    if (bind(qgc_socket_fd_, (struct sockaddr *)&local_qgc_addr_, local_qgc_addr_len_) < 0) {
-      ignerr << "QGC UDP bind failed: " << strerror(errno) << ", aborting\n";
-      abort();
-    }
-
-    if ((sdk_socket_fd_ = socket(AF_INET, SOCK_DGRAM, 0)) < 0) {
-      ignerr << "Creating SDK UDP socket failed: " << strerror(errno) << ", aborting\n";
-      abort();
-    }
-
-    if (bind(sdk_socket_fd_, (struct sockaddr *)&local_sdk_addr_, local_sdk_addr_len_) < 0) {
-      ignerr << "SDK UDP bind failed: " << strerror(errno) << ", aborting\n";
-      abort();
-    }
-
+    int sdk_udp_port = _sdf->Get<int>("sdk_udp_port");
+    mavlink_interface_->SetSdkUdpPort(sdk_udp_port);
   }
 
   if (serial_enabled_) {
     // Set up serial interface
     if(_sdf->HasElement("serialDevice"))
     {
-      device_ = _sdf->Get<std::string>("serialDevice");
+      std::string device = _sdf->Get<std::string>("serialDevice");
     }
 
     if (_sdf->HasElement("baudRate")) {
-      baudrate_ = _sdf->Get<int>("baudRate");
+      int baudrate = _sdf->Get<int>("baudRate");
     }
-    io_service.post(std::bind(&GazeboMavlinkInterface::do_read, this));
-
-    // run io_service for async io
-    io_thread = std::thread([this] () {
-      io_service.run();
-    });
-    open();
-
-  } else {
-    memset((char *)&remote_simulator_addr_, 0, sizeof(remote_simulator_addr_));
-    remote_simulator_addr_.sin_family = AF_INET;
-    remote_simulator_addr_len_ = sizeof(remote_simulator_addr_);
-
-    memset((char *)&local_simulator_addr_, 0, sizeof(local_simulator_addr_));
-    local_simulator_addr_.sin_family = AF_INET;
-    local_simulator_addr_len_ = sizeof(local_simulator_addr_);
-
-    if (use_tcp_) {
-
-      local_simulator_addr_.sin_addr.s_addr = htonl(mavlink_addr_);
-      local_simulator_addr_.sin_port = htons(mavlink_tcp_port_);
-
-      if ((simulator_socket_fd_ = socket(AF_INET, SOCK_STREAM, 0)) < 0) {
-        ignerr << "Creating TCP socket failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      int yes = 1;
-      int result = setsockopt(simulator_socket_fd_, IPPROTO_TCP, TCP_NODELAY, &yes, sizeof(yes));
-      if (result != 0) {
-        ignerr << "setsockopt failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      struct linger nolinger {};
-      nolinger.l_onoff = 1;
-      nolinger.l_linger = 0;
-
-      result = setsockopt(simulator_socket_fd_, SOL_SOCKET, SO_LINGER, &nolinger, sizeof(nolinger));
-      if (result != 0) {
-        ignerr << "setsockopt failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      // The socket reuse is necessary for reconnecting to the same address
-      // if the socket does not close but gets stuck in TIME_WAIT. This can happen
-      // if the server is suddenly closed, for example, if the robot is deleted in gazebo.
-      int socket_reuse = 1;
-      result = setsockopt(simulator_socket_fd_, SOL_SOCKET, SO_REUSEADDR, &socket_reuse, sizeof(socket_reuse));
-      if (result != 0) {
-        ignerr << "setsockopt failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      // Same as above but for a given port
-      result = setsockopt(simulator_socket_fd_, SOL_SOCKET, SO_REUSEPORT, &socket_reuse, sizeof(socket_reuse));
-      if (result != 0) {
-        ignerr << "setsockopt failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      // set socket to non-blocking
-      result = fcntl(simulator_socket_fd_, F_SETFL, O_NONBLOCK);
-      if (result == -1) {
-        ignerr << "setting socket to non-blocking failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      if (bind(simulator_socket_fd_, (struct sockaddr *)&local_simulator_addr_, local_simulator_addr_len_) < 0) {
-        ignerr << "bind failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      errno = 0;
-      if (listen(simulator_socket_fd_, 0) < 0) {
-        ignerr << "listen failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      memset(fds_, 0, sizeof(fds_));
-      fds_[LISTEN_FD].fd = simulator_socket_fd_;
-      fds_[LISTEN_FD].events = POLLIN; // only listens for new connections on tcp
-
-    } else {
-      remote_simulator_addr_.sin_addr.s_addr = mavlink_addr_;
-      remote_simulator_addr_.sin_port = htons(mavlink_udp_port_);
-
-      local_simulator_addr_.sin_addr.s_addr = htonl(INADDR_ANY);
-      local_simulator_addr_.sin_port = htons(0);
-
-      if ((simulator_socket_fd_ = socket(AF_INET, SOCK_DGRAM, 0)) < 0) {
-        ignerr << "Creating UDP socket failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      // set socket to non-blocking
-      int result = fcntl(simulator_socket_fd_, F_SETFL, O_NONBLOCK);
-      if (result == -1) {
-        ignerr << "setting socket to non-blocking failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      if (bind(simulator_socket_fd_, (struct sockaddr *)&local_simulator_addr_, local_simulator_addr_len_) < 0) {
-        ignerr << "bind failed: " << strerror(errno) << ", aborting\n";
-        abort();
-      }
-
-      memset(fds_, 0, sizeof(fds_));
-      fds_[CONNECTION_FD].fd = simulator_socket_fd_;
-      fds_[CONNECTION_FD].events = POLLIN | POLLOUT; // read/write
-    }
-  }
-
-  if(_sdf->HasElement("vehicle_is_tailsitter"))
-  {
-    vehicle_is_tailsitter_ = _sdf->Get<bool>("vehicle_is_tailsitter");
   }
 
   if(_sdf->HasElement("send_vision_estimation"))
@@ -629,9 +440,7 @@ void GazeboMavlinkInterface::Configure(const ignition::gazebo::Entity &_entity,
   else {
     ignerr << "Unkown protocol version! Using v" << protocol_version_ << "by default \n";
   }
-
-  standard_normal_distribution_ = std::normal_distribution<float>(0.0f, 1.0f);
-
+  mavlink_interface_->Load();
 }
 
 void GazeboMavlinkInterface::PreUpdate(const ignition::gazebo::UpdateInfo &_info,
@@ -651,13 +460,20 @@ void GazeboMavlinkInterface::PreUpdate(const ignition::gazebo::UpdateInfo &_info
     return;
   }
 
-  double dt = std::chrono::duration<double>(_info.dt).count();
+// #if GAZEBO_MAJOR_VERSION >= 9
+//   common::Time current_time = world_->SimTime();
+// #else
+//   common::Time current_time = world_->GetSimTime();
+// #endif
+//   double dt = (current_time - last_time_).Double();
+  double dt;
 
-  close_conn_ = false;
+  bool close_conn_ = false;
+
   if (hil_mode_) {
-    pollFromQgcAndSdk();
+    mavlink_interface_->pollFromQgcAndSdk();
   } else {
-    pollForMAVLinkMessages();
+    mavlink_interface_->pollForMAVLinkMessages();
   }
 
   // Always send Gyro and Accel data at full rate (= sim update rate)
@@ -667,7 +483,7 @@ void GazeboMavlinkInterface::PreUpdate(const ignition::gazebo::UpdateInfo &_info
   SendGroundTruth();
 
   if (close_conn_) { // close connection if required
-    close();
+    mavlink_interface_->close();
   }
 
   handle_control(dt);
@@ -693,33 +509,6 @@ void GazeboMavlinkInterface::PreUpdate(const ignition::gazebo::UpdateInfo &_info
 
 void GazeboMavlinkInterface::PostUpdate(const ignition::gazebo::UpdateInfo &_info,
     const ignition::gazebo::EntityComponentManager &_ecm) {
-}
-
-void GazeboMavlinkInterface::forward_mavlink_message(const mavlink_message_t *message)
-{
-  if (gotSigInt_) {
-    return;
-  }
-
-  uint8_t buffer[MAVLINK_MAX_PACKET_LEN];
-  int packetlen = mavlink_msg_to_send_buffer(buffer, message);
-  ssize_t len;
-  if (qgc_socket_fd_ > 0) {
-    len = sendto(qgc_socket_fd_, buffer, packetlen, 0, (struct sockaddr *)&remote_qgc_addr_, remote_qgc_addr_len_);
-
-    if (len <= 0)
-    {
-      ignerr << "Failed sending mavlink message to QGC: " << strerror(errno) << "\n";
-    }
-  }
-
-  if (sdk_socket_fd_ > 0) {
-    len = sendto(sdk_socket_fd_, buffer, packetlen, 0, (struct sockaddr *)&remote_sdk_addr_, remote_sdk_addr_len_);
-    if (len <= 0)
-    {
-      ignerr << "Failed sending mavlink message to SDK: " << strerror(errno) << "\n";
-    }
-  }
 }
 
 void GazeboMavlinkInterface::ImuCallback(const ignition::msgs::IMU &_msg) {
@@ -796,118 +585,10 @@ void GazeboMavlinkInterface::SendSensorMessages() {
     mag_updated_ = false;
   }
 
-  // send only baro data
-  if (baro_updated_) {
-    sensor_msg.temperature = temperature_;
-    sensor_msg.abs_pressure = abs_pressure_;
-    sensor_msg.pressure_alt = pressure_alt_;
-    // sensor_msg.fields_updated = sensor_msg.fields_updated | SensorSource::BARO;
-
-    baro_updated_ = false;
-  }
-
-  // send only diff pressure data
-  if (diff_press_updated_) {
-    const float temperature_msl = 288.0f; // temperature at MSL (Kelvin)
-    float temperature_local = sensor_msg.temperature + 273.0f;
-    const float density_ratio = powf((temperature_msl/temperature_local) , 4.256f);
-    float rho = 1.225f / density_ratio;
-
-    // Let's use a rough guess of 0.01 hPa as the standard devitiation which roughly yields
-    // about +/- 1 m/s noise.
-    const float diff_pressure_stddev = 0.01f;
-    const float diff_pressure_noise = standard_normal_distribution_(random_generator_) * diff_pressure_stddev;
-
-// #if GAZEBO_MAJOR_VERSION >= 9
-    ignition::math::Vector3d vel_b; //TODO: How do you get body velocity?
-//     ignition::math::Vector3d vel_b = q_br.RotateVector(model_->RelativeLinearVel());
-// #else
-//     ignition::math::Vector3d vel_b = q_br.RotateVector(ignitionFromGazeboMath(model_->GetRelativeLinearVel()));
-// #endif
-
-    // calculate differential pressure in hPa
-    // if vehicle is a tailsitter the airspeed axis is different (z points from nose to tail)
-    if (vehicle_is_tailsitter_) {
-      sensor_msg.diff_pressure = 0.005f * rho * vel_b.Z() * vel_b.Z() + diff_pressure_noise;
-    } else {
-      sensor_msg.diff_pressure = 0.005f * rho * vel_b.X() * vel_b.X() + diff_pressure_noise;
-    }
-    // sensor_msg.fields_updated = sensor_msg.fields_updated | SensorSource::DIFF_PRESS;
-
-    diff_press_updated_ = false;
-  }
-
   if (!hil_mode_ || (hil_mode_ && !hil_state_level_)) {
     mavlink_message_t msg;
     mavlink_msg_hil_sensor_encode_chan(1, 200, MAVLINK_COMM_0, &msg, &sensor_msg);
-    send_mavlink_message(&msg);
-  }
-}
-
-void GazeboMavlinkInterface::send_mavlink_message(const mavlink_message_t *message)
-{
-  assert(message != nullptr);
-
-  if (gotSigInt_ || close_conn_) {
-    return;
-  }
-
-  if (serial_enabled_) {
-
-    if (!is_open()) {
-      ignerr << "Serial port closed! \n";
-      return;
-    }
-
-    {
-      std::lock_guard<std::recursive_mutex> lock(mutex);
-
-      if (tx_q.size() >= MAX_TXQ_SIZE) {
-        ignwarn << "Tx queue overflow\n";
-      }
-      tx_q.emplace_back(message);
-    }
-    io_service.post(std::bind(&GazeboMavlinkInterface::do_write, this, true));
-
-  } else {
-    uint8_t buffer[MAVLINK_MAX_PACKET_LEN];
-    int packetlen = mavlink_msg_to_send_buffer(buffer, message);
-
-    if (fds_[CONNECTION_FD].fd > 0) {
-      int timeout_ms = (received_first_actuator_ && enable_lockstep_) ? 1000 : 0;
-      int ret = ::poll(&fds_[0], N_FDS, timeout_ms);
-
-      if (ret < 0) {
-        ignerr << "poll error: " << strerror(errno) << "\n";
-        return;
-      }
-
-      if (ret == 0 && timeout_ms > 0) {
-        ignerr << "poll timeout\n";
-        return;
-      }
-
-      if (!(fds_[CONNECTION_FD].revents & POLLOUT)) {
-        ignerr << "invalid events at fd:" << fds_[CONNECTION_FD].revents << "\n";
-        return;
-      }
-
-      size_t len;
-      if (use_tcp_) {
-        len = send(fds_[CONNECTION_FD].fd, buffer, packetlen, 0);
-      } else {
-        len = sendto(fds_[CONNECTION_FD].fd, buffer, packetlen, 0, (struct sockaddr *)&remote_simulator_addr_, remote_simulator_addr_len_);
-      }
-      if (len < 0) {
-        ignerr << "Failed sending mavlink message: " << strerror(errno) << "\n";
-        if (errno == ECONNRESET || errno == EPIPE) {
-          if (use_tcp_) { // udp socket remains alive
-            ignerr << "Closing connection." << "\n";
-            close_conn_ = true;
-          }
-        }
-      }
-    }
+    mavlink_interface_->send_mavlink_message(&msg);
   }
 }
 
@@ -973,180 +654,7 @@ void GazeboMavlinkInterface::SendGroundTruth()
   if (!hil_mode_ || (hil_mode_ && hil_state_level_)) {
     mavlink_message_t msg;
     mavlink_msg_hil_state_quaternion_encode_chan(1, 200, MAVLINK_COMM_0, &msg, &hil_state_quat);
-    send_mavlink_message(&msg);
-  }
-}
-
-void GazeboMavlinkInterface::pollForMAVLinkMessages()
-{
-  if (gotSigInt_) {
-    return;
-  }
-
-  bool received_actuator = false;
-
-  do {
-    int timeout_ms = (received_first_actuator_ && enable_lockstep_) ? 1000 : 0;
-    int ret = ::poll(&fds_[0], N_FDS, timeout_ms);
-
-    if (ret < 0) {
-      ignerr << "poll error: " << strerror(errno) << "\n";
-      return;
-    }
-
-    if (ret == 0 && timeout_ms > 0) {
-      ignerr << "poll timeout\n";
-      return;
-    }
-
-    for (int i = 0; i < N_FDS; i++) {
-      if(fds_[i].revents == 0) {
-        continue;
-      }
-
-      if (!(fds_[i].revents & POLLIN)) {
-        continue;
-      }
-
-      if (i == LISTEN_FD) { // if event is raised on the listening socket
-        acceptConnections();
-      } else { // if event is raised on connection socket
-        int ret = recvfrom(fds_[i].fd, _buf, sizeof(_buf), 0, (struct sockaddr *)&remote_simulator_addr_, &remote_simulator_addr_len_);
-        if (ret < 0) {
-          // all data is read if EWOULDBLOCK is raised
-          if (errno != EWOULDBLOCK) { // disconnected from client
-            ignerr << "recvfrom error: " << strerror(errno) << "\n";
-          }
-          continue;
-        }
-
-        // client closed the connection orderly, only makes sense on tcp
-        if (use_tcp_ && ret == 0) {
-          ignerr << "Connection closed by client." << "\n";
-          close_conn_ = true;
-          continue;
-        }
-
-        // data received
-        int len = ret;
-        mavlink_message_t msg;
-        mavlink_status_t status;
-        for (unsigned i = 0; i < len; ++i) {
-          if (mavlink_parse_char(MAVLINK_COMM_0, _buf[i], &msg, &status)) {
-            if (hil_mode_) {
-              send_mavlink_message(&msg);
-            }
-            handle_message(&msg, received_actuator);
-          }
-        }
-      }
-    }
-  } while (!close_conn_ && received_first_actuator_ && !received_actuator && enable_lockstep_ && IsRunning() && !gotSigInt_);
-}
-
-void GazeboMavlinkInterface::acceptConnections()
-{
-  if (fds_[CONNECTION_FD].fd > 0) {
-    return;
-  }
-
-  // accepting incoming connections on listen fd
-  int ret =
-    accept(fds_[LISTEN_FD].fd, (struct sockaddr *)&remote_simulator_addr_, &remote_simulator_addr_len_);
-
-  if (ret < 0) {
-    if (errno != EWOULDBLOCK) {
-      ignerr << "accept error: " << strerror(errno) << "\n";
-    }
-    return;
-  }
-
-  // assign socket to connection descriptor on success
-  fds_[CONNECTION_FD].fd = ret; // socket is replaced with latest connection
-  fds_[CONNECTION_FD].events = POLLIN | POLLOUT; // read/write
-}
-
-void GazeboMavlinkInterface::pollFromQgcAndSdk()
-{
-  struct pollfd fds[2] = {};
-  fds[0].fd = qgc_socket_fd_;
-  fds[0].events = POLLIN;
-  fds[1].fd = sdk_socket_fd_;
-  fds[1].events = POLLIN;
-
-  const int timeout_ms = 0;
-
-  int ret = ::poll(&fds[0], 2, timeout_ms);
-
-  if (ret < 0) {
-    ignerr << "poll error: " << strerror(errno) << "\n";
-    return;
-  }
-
-  if (fds[0].revents & POLLIN) {
-    int len = recvfrom(qgc_socket_fd_, _buf, sizeof(_buf), 0, (struct sockaddr *)&remote_qgc_addr_, &remote_qgc_addr_len_);
-
-    if (len > 0) {
-      mavlink_message_t msg;
-      mavlink_status_t status;
-      for (unsigned i = 0; i < len; ++i) {
-        if (mavlink_parse_char(MAVLINK_COMM_1, _buf[i], &msg, &status)) {
-          // forward message from QGC to serial
-          send_mavlink_message(&msg);
-        }
-      }
-    }
-  }
-
-  if (fds[1].revents & POLLIN) {
-    int len = recvfrom(sdk_socket_fd_, _buf, sizeof(_buf), 0, (struct sockaddr *)&remote_sdk_addr_, &remote_sdk_addr_len_);
-
-    if (len > 0) {
-      mavlink_message_t msg;
-      mavlink_status_t status;
-      for (unsigned i = 0; i < len; ++i) {
-        if (mavlink_parse_char(MAVLINK_COMM_2, _buf[i], &msg, &status)) {
-          // forward message from SDK to serial
-          send_mavlink_message(&msg);
-        }
-      }
-    }
-  }
-}
-
-void GazeboMavlinkInterface::handle_message(mavlink_message_t *msg, bool &received_actuator)
-{
-  switch (msg->msgid) {
-  case MAVLINK_MSG_ID_HIL_ACTUATOR_CONTROLS:
-    mavlink_hil_actuator_controls_t controls;
-    mavlink_msg_hil_actuator_controls_decode(msg, &controls);
-
-    bool armed = (controls.mode & MAV_MODE_FLAG_SAFETY_ARMED);
-
-// #if GAZEBO_MAJOR_VERSION >= 9
-//     last_actuator_time_ = world_->SimTime();
-// #else
-//     last_actuator_time_ = world_->GetSimTime();
-// #endif
-
-    for (unsigned i = 0; i < n_out_max; i++) {
-      input_index_[i] = i;
-    }
-
-    // set rotor speeds, controller targets
-    input_reference_.resize(n_out_max);
-    for (int i = 0; i < input_reference_.size(); i++) {
-      if (armed) {
-        input_reference_[i] = (controls.controls[input_index_[i]] + input_offset_[i])
-            * input_scaling_[i] + zero_position_armed_[i];
-      } else {
-        input_reference_[i] = zero_position_disarmed_[i];
-      }
-    }
-
-    received_actuator = true;
-    received_first_actuator_ = true;
-    break;
+    mavlink_interface_->send_mavlink_message(&msg);
   }
 }
 
@@ -1219,130 +727,4 @@ bool GazeboMavlinkInterface::IsRunning()
 //     return world_->GetRunning();
 // #endif
   return true; //TODO;
-}
-
-void GazeboMavlinkInterface::open() {
-  try{
-    serial_dev.open(device_);
-    serial_dev.set_option(boost::asio::serial_port_base::baud_rate(baudrate_));
-    serial_dev.set_option(boost::asio::serial_port_base::character_size(8));
-    serial_dev.set_option(boost::asio::serial_port_base::parity(boost::asio::serial_port_base::parity::none));
-    serial_dev.set_option(boost::asio::serial_port_base::stop_bits(boost::asio::serial_port_base::stop_bits::one));
-    serial_dev.set_option(boost::asio::serial_port_base::flow_control(boost::asio::serial_port_base::flow_control::none));
-    igndbg << "Opened serial device " << device_ << "\n";
-  }
-  catch (boost::system::system_error &err) {
-    ignerr <<"Error opening serial device: " << err.what() << "\n";
-  }
-}
-
-void GazeboMavlinkInterface::close()
-{
-  if(serial_enabled_) {
-    ::close(qgc_socket_fd_);
-    ::close(sdk_socket_fd_);
-
-    std::lock_guard<std::recursive_mutex> lock(mutex);
-    if (!is_open())
-      return;
-
-    io_service.stop();
-    serial_dev.close();
-
-    if (io_thread.joinable())
-      io_thread.join();
-
-  } else {
-
-    ::close(fds_[CONNECTION_FD].fd);
-    fds_[CONNECTION_FD] = { 0, 0, 0 };
-    fds_[CONNECTION_FD].fd = -1;
-
-    received_first_actuator_ = false;
-
-  }
-}
-
-void GazeboMavlinkInterface::do_read(void)
-{
-  serial_dev.async_read_some(boost::asio::buffer(rx_buf), boost::bind(
-      &GazeboMavlinkInterface::parse_buffer, this, boost::asio::placeholders::error, boost::asio::placeholders::bytes_transferred
-      )
-  );
-}
-
-// Based on MAVConnInterface::parse_buffer in MAVROS
-void GazeboMavlinkInterface::parse_buffer(const boost::system::error_code& err, std::size_t bytes_t){
-  mavlink_status_t status;
-  mavlink_message_t message;
-  uint8_t *buf = this->rx_buf.data();
-
-  assert(rx_buf.size() >= bytes_t);
-
-  for(; bytes_t > 0; bytes_t--)
-  {
-    auto c = *buf++;
-
-    auto msg_received = static_cast<Framing>(mavlink_frame_char_buffer(&m_buffer, &m_status, c, &message, &status));
-    if (msg_received == Framing::bad_crc || msg_received == Framing::bad_signature) {
-      _mav_parse_error(&m_status);
-      m_status.msg_received = MAVLINK_FRAMING_INCOMPLETE;
-      m_status.parse_state = MAVLINK_PARSE_STATE_IDLE;
-      if (c == MAVLINK_STX) {
-        m_status.parse_state = MAVLINK_PARSE_STATE_GOT_STX;
-        m_buffer.len = 0;
-        mavlink_start_checksum(&m_buffer);
-      }
-    }
-
-    if (msg_received != Framing::incomplete) {
-      if (hil_mode_) {
-        forward_mavlink_message(&message);
-      }
-      bool not_used;
-      handle_message(&message, not_used);
-    }
-  }
-  do_read();
-}
-
-void GazeboMavlinkInterface::do_write(bool check_tx_state){
-  if (check_tx_state && tx_in_progress)
-    return;
-
-  std::lock_guard<std::recursive_mutex> lock(mutex);
-  if (tx_q.empty())
-    return;
-
-  tx_in_progress = true;
-  auto &buf_ref = tx_q.front();
-
-  serial_dev.async_write_some(
-    boost::asio::buffer(buf_ref.dpos(), buf_ref.nbytes()), [this, &buf_ref] (boost::system::error_code error,   size_t bytes_transferred)
-    {
-      assert(bytes_transferred <= buf_ref.len);
-      if(error) {
-        ignerr << "Serial error: " << error.message() << "\n";
-      return;
-      }
-
-    std::lock_guard<std::recursive_mutex> lock(mutex);
-
-    if (tx_q.empty()) {
-      tx_in_progress = false;
-      return;
-    }
-
-    buf_ref.pos += bytes_transferred;
-    if (buf_ref.nbytes() == 0) {
-      tx_q.pop_front();
-    }
-
-    if (!tx_q.empty()) {
-      do_write(false);
-    }
-    else {
-      tx_in_progress = false;
-    }
-  });
 }
